+++ source/src/Data/MissionData/View/EnhancedBattleTestViews.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using EnhancedBattleTest.UI.MissionUI;
+using SandBox.View.Missions;
 using TaleWorlds.MountAndBlade;
 using TaleWorlds.MountAndBlade.LegacyGUI.Missions;
 using TaleWorlds.MountAndBlade.View;
@@ -11,7 +12,7 @@ namespace EnhancedBattleTest.Data.MissionData.View
     public class EnhancedBattleTestViews
     {
         [ViewMethod("EnhancedBattleTestFieldBattle")]
-        public static MissionView[] OpenMPBattleMission(Mission mission)
+        public static MissionView[] OpenBattleMission(Mission mission)
         {
             return new List<MissionView>
             {
@@ -21,8 +22,10 @@ namespace EnhancedBattleTest.Data.MissionData.View
                 ViewCreator.CreateOptionsUIHandler(),
                 ViewCreator.CreateMissionOrderUIHandler(),
                 new OrderTroopPlacer(),
+                new MissionSingleplayerUIHandler(),
                 ViewCreator.CreateMissionAgentStatusUIHandler(mission),
                 ViewCreator.CreateMissionMainAgentEquipmentController(mission),
+                ViewCreator.CreateMissionMainAgentCheerControllerView(mission),
                 new MusicBattleMissionView(false),
                 ViewCreator.CreateMissionBoundaryCrossingView(),
                 new MissionBoundaryWallView(),
@@ -36,7 +39,7 @@ namespace EnhancedBattleTest.Data.MissionData.View
             }.ToArray();
         }
         [ViewMethod("EnhancedBattleTestSiegeBattle")]
-        public static MissionView[] OpenMPSiegeBattleMission(Mission mission)
+        public static MissionView[] OpenSiegeBattleMission(Mission mission)
         {
             MissionView missionOrderUiHandler = ViewCreator.CreateMissionOrderUIHandler();
             ISiegeDeploymentView siegeDeploymentView = missionOrderUiHandler as ISiegeDeploymentView;
@@ -48,8 +51,10 @@ namespace EnhancedBattleTest.Data.MissionData.View
                 ViewCreator.CreateOptionsUIHandler(),
                 missionOrderUiHandler,
                 new OrderTroopPlacer(),
+                new MissionSingleplayerUIHandler(),
                 ViewCreator.CreateMissionAgentStatusUIHandler(mission),
                 ViewCreator.CreateMissionMainAgentEquipmentController(mission),
+                ViewCreator.CreateMissionMainAgentCheerControllerView(mission),
                 new MusicBattleMissionView(true),
                 new SiegeMissionView(),
                 new MissionEntitySelectionUIHandler(siegeDeploymentView.OnEntitySelection,