+++ source/src/SinglePlayer/Data/MissionData/SPCombatant.cs
@@ -42,6 +42,13 @@ namespace EnhancedBattleTest.SinglePlayer.Data.MissionData
         public static SPCombatant CreateParty(PartyBase party, BattleSideEnum side, BasicCultureObject culture,
             TeamConfig teamConfig, bool isPlayerTeam)
         {
+            party.Owner = null;
+            if (teamConfig.HasGeneral)
+            {
+                var characterObject = (teamConfig.General as SPCharacterConfig)?.ActualCharacterObject;
+                if (characterObject?.IsHero ?? false)
+                    party.Owner = characterObject.HeroObject;
+            }
             Utility.FillPartyMembers(party, side, culture, teamConfig, isPlayerTeam);
             bool isAttacker = side == BattleSideEnum.Attacker;
             var combatant = new SPCombatant(party, side, teamConfig.TacticLevel, culture,